chore: move new sorted set fetch response type to new requests dir
This is just to keep the new files separated from the old files.
The sorted_set_fetch response definition was inside the old `responses`
dir, which i'm expecting us to delete before too long. This just moves
it over to the new directory structure. We can reorganize a bit further
after we delete the old code if necessary.
