+++ src/cache_client.rs
@@ -17,6 +17,7 @@ use crate::requests::cache::sorted_set::sorted_set_fetch_by_rank::{
     SortOrder, SortedSetFetchByRankRequest,
 };
 use crate::requests::cache::sorted_set::sorted_set_fetch_by_score::SortedSetFetchByScoreRequest;
+use crate::requests::cache::sorted_set::sorted_set_fetch_response::SortedSetFetch;
 use crate::requests::cache::sorted_set::sorted_set_put_element::{
     SortedSetPutElement, SortedSetPutElementRequest,
 };
@@ -24,7 +25,6 @@ use crate::requests::cache::sorted_set::sorted_set_put_elements::{
     SortedSetPutElements, SortedSetPutElementsRequest,
 };
 use crate::requests::cache::MomentoRequest;
-use crate::response::cache::sorted_set_fetch::SortedSetFetch;
 
 use crate::cache_client_builder::{CacheClientBuilder, NeedsDefaultTtl};
 use crate::{utils, IntoBytes, MomentoResult};
@@ -487,7 +487,7 @@ impl CacheClient {
     /// # use momento_test_util::create_doctest_cache_client;
     /// # tokio_test::block_on(async {
     /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_rank::SortOrder;
-    /// use momento::response::cache::sorted_set_fetch::SortedSetFetch;
+    /// use momento::requests::cache::sorted_set::sorted_set_fetch_response::SortedSetFetch;
     /// # let (cache_client, cache_name) = create_doctest_cache_client();
     /// let sorted_set_name = "sorted_set";
     ///
@@ -522,7 +522,7 @@ impl CacheClient {
     /// # tokio_test::block_on(async {
     /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_rank::SortOrder;
     /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_rank::SortedSetFetchByRankRequest;
-    /// use momento::response::cache::sorted_set_fetch::SortedSetFetch;
+    /// use momento::requests::cache::sorted_set::sorted_set_fetch_response::SortedSetFetch;
     /// # let (cache_client, cache_name) = create_doctest_cache_client();
     /// let sorted_set_name = "sorted_set";
     ///
@@ -585,7 +585,7 @@ impl CacheClient {
     /// # use momento_test_util::create_doctest_cache_client;
     /// # tokio_test::block_on(async {
     /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_rank::SortOrder;
-    /// use momento::response::cache::sorted_set_fetch::SortedSetFetch;
+    /// use momento::requests::cache::sorted_set::sorted_set_fetch_response::SortedSetFetch;
     /// # let (cache_client, cache_name) = create_doctest_cache_client();
     /// let sorted_set_name = "sorted_set";
     ///
@@ -620,7 +620,7 @@ impl CacheClient {
     /// # tokio_test::block_on(async {
     /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_rank::SortOrder;
     /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_score::SortedSetFetchByScoreRequest;
-    /// use momento::response::cache::sorted_set_fetch::SortedSetFetch;
+    /// use momento::requests::cache::sorted_set::sorted_set_fetch_response::SortedSetFetch;
     /// # let (cache_client, cache_name) = create_doctest_cache_client();
     /// let sorted_set_name = "sorted_set";
     ///
@@ -667,7 +667,7 @@ impl CacheClient {
     /// use momento::config::configurations;
     /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_rank::SortedSetFetchByRankRequest;
     /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_rank::SortOrder;
-    /// use momento::response::cache::sorted_set_fetch::SortedSetFetch;
+    /// use momento::requests::cache::sorted_set::sorted_set_fetch_response::SortedSetFetch;
     ///
     /// let credential_provider = CredentialProvider::from_env_var("MOMENTO_API_KEY".to_string())?;
     /// let cache_name = "cache";
