+++ src/requests/cache/sorted_set/mod.rs
@@ -1,4 +1,5 @@
 pub mod sorted_set_fetch_by_rank;
 pub mod sorted_set_fetch_by_score;
+pub mod sorted_set_fetch_response;
 pub mod sorted_set_put_element;
 pub mod sorted_set_put_elements;
