+++ src/requests/cache/sorted_set/sorted_set_fetch_by_rank.rs
@@ -1,8 +1,8 @@
 use momento_protos::cache_client::sorted_set_fetch_request::{by_index, ByIndex, Range};
 use momento_protos::cache_client::{SortedSetFetchRequest, Unbounded};
 
+use crate::requests::cache::sorted_set::sorted_set_fetch_response::SortedSetFetch;
 use crate::requests::cache::MomentoRequest;
-use crate::response::cache::sorted_set_fetch::SortedSetFetch;
 use crate::simple_cache_client::prep_request_with_timeout;
 use crate::{CacheClient, IntoBytes, MomentoResult};
 
@@ -38,7 +38,7 @@ pub enum SortOrder {
 /// # tokio_test::block_on(async {
 /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_rank::SortOrder;
 /// use momento::requests::cache::sorted_set::sorted_set_fetch_by_rank::SortedSetFetchByRankRequest;
-/// use momento::response::cache::sorted_set_fetch::SortedSetFetch;
+/// use momento::requests::cache::sorted_set::sorted_set_fetch_response::SortedSetFetch;
 /// # let (cache_client, cache_name) = create_doctest_cache_client();
 /// let sorted_set_name = "sorted_set";
 ///
