--- src/response/cache/mod.rs
@@ -1,1 +1,0 @@
-pub mod sorted_set_fetch;
